pr feedback &:code cleanup
